Delete binding from db when binding request to service broker fails
When the binding request to the service broker fails (e.g. times out),
orphan mitigation was already triggered (i.e. unbind request sent to the
service broker) but the database entry was kept. This led to errors when
calling 'cf bind-service' a second time (App .. is already bound to
service instance ..).

With this change we remove the binding from the database when the
initial binding request to the service broker fails. Thus we also
changed an already existing test.

Co-authored-by: Katharina Przybill <katharina.przybill@sap.com>
Co-authored-by: Johannes Haass <johannes.haass@sap.com>