Throw a new error when sb binding is in deletion of deletion failed
When binding is in state delete_in_progess or delete failed we want an
error different from "binding already exists" to be thrown. Fistly this
might not be true since in the meantime orphan mitigation might already
deleted that binding. Furthermore the CF CLI interprets this(rightfully) as
success since this message tells it that what CF CLI wanted to to is
already accomplished(creating that binding) when in may not exist.
So CF CLI will now correctly fail on create-credential-binding when a binding exists that
is in a deletion_failed por deletion_in_progress state.
Same applies for cf cli command create-credential-key, they share the
same endpoint but have other backend logic. Both cases, bindings and
keys where improoved.