Prevent inconsistencies when cancelling service creations
Although the Open Service Broker API specification allows for delete
requests being processed in parallel to create requests (which should
semantically be treated as a cancel operation), it is safer to only
allow this when the broker initially responded to the creation request.

Thus a new 'create initial' state is used to distinguish between
requests that have not been responded to and operations that are being
processed asynchronously by the broker.

The 'initial' state is set in:
- ServiceCredentialBindingAppCreate
- ServiceCredentialBindingKeyCreate
- ServiceRouteBindingCreate
- ServiceInstanceCreateManaged

When a create request is stuck (i.e. the request to the broker times out
after 60 seconds), it is no longer possible to issue a delete request in
parallel. If the create request has reached the broker and is being
processed asynchronously, a delete request can be issued by the client
to cancel the operation.