Creates metrics endpoint for use by prom-scraper
Notable changes in this commit:
* Adds an `/internal/v4/metrics` endpoint which serves Prometheus-style
metrics.
* Adds function to Prometheus::Client::Registry to re-create the
library's metrics registry. This is used by the PrometheusClient class's
tests to wipe out the metrics registry, which is global state. Global
state is the #1 cause of test pollution and this change seemed to be the
most straightforward way to eliminate some order-dependent test failures.
NOTE: This function definition is only in the
`prometheus_updater_spec.rb` file, which we believe is not shipped in
capi-release, so the odds of actually writing production code that
depends on this test-only functionality are pretty slim.

Co-authored-by: David Alvarado <alvaradoda@vmware.com>
Co-authored-by: Kenneth Lakin <klakin@vmware.com>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Michael Oleske <moleske@pivotal.io>