Implements sharing a route across spaces
This commit adds 'POST /v3/routes/:guid/relationships/shared_spaces'.
The endpoint's behavior currently matches service instance sharing behavior.
Route sharing does not handle isolation segments.

This endpoint is currently experimental.

This commit also makes minor changes to the error message transmitted
when you fail to share a Service Instance. The existing message didn't
tell you what instance failed to share, or the space that caused the
failure,  which are both nice things for an operator to know.

Here is a note about some implementation detail for future programmers:

`Route.shared?` does not make use of the class's `shared_spaces`
member variable. This is because `shared_spaces` is an `Array`, which
indicates that -behind the scenes- Sequel will first fetch all of the
data for all of the matching objects in the association, wrap that in
an array, and then hand it to us. Given that we only want to know if
there are more than zero items in that array, this could be a huge
waste on Foundations that have a route shared between many spaces.

So, we explicitly search for `Spaces` associated with the `Route` and tell
the ORM to tell us if there is more than one. The Sequel docs claim that
using `empty?` on a `Dataset` object implicitly adds in a `LIMIT 1`
to the query that the ORM generates, so the code as written should only
ever send us one row at most.

[#181757727]

Co-authored-by: Alex Rocha <alexr1@vmware.com>
Co-authored-by: Merric de Launey <mdelauney@pivotal.io>
Co-authored-by: Kenneth Lakin <klakin@vmware.com>