Check arguments encoding before dispatching to controller (v2)
Whereas Rails (/v3) validates the encoding of path parameters and
responds with the error message "Invalid path parameters: Invalid
encoding for parameter: ..." in case of an invalid encoding, Sinatra
(/v2) simply passes the arguments on to the controller.

With this change all arguments that implement 'valid_encoding?' are
checked and if this validation fails an InvalidRequest error is
returned.