Improve performance of service plan / offering list fetching
Using UNION instead of multiple ORed conditions (see [1]) is
significantly faster for larger databases.

The BaseServiceListFetcher has been rewritten to perform three different
sub-queries in order to retrieve plans / offerings that are
 - readable via the plan org (aka service plan visibilities),
 - readable via the broker space (i.e. space-scoped service brokers),
 - publicly readable.

Additional filters are pushed down to these specific sub-queries instead
of being applied to the overall result set.

Due to a MySQL issue (see [2]) boolean fields need an explicit
conversion as the original type is not maintained for a UNION result.

[1] http://blog.rhodiumtoad.org.uk/2017/01/22/performance-issues-with-ored-conditions/
[2] https://bugs.mysql.com/bug.php?id=30886

Co-authored-by: Florian Braun <5863788+FloThinksPi@users.noreply.github.com>
Co-authored-by: Stephan Merker <stephan.merker@sap.com>