Fix old migration files and add missing indices
Sequel 5.56.0 added support for the :index option as part of add_column
(see [1]). Until now this was not supported, but used in 4 places
(without effect, i.e. indices had not been created).

In order to have an identical database schema for newly created as well
as already exiting databases, the previously unsupported :index option
is removed from the old migration files and a new migration is added
that creates the 3 missing indices (one of the columns had been removed
in the meantime).

[1] https://github.com/jeremyevans/sequel/blob/5.56.0/CHANGELOG#L3