Add separate check for active org
The checks if an org is suspended and the user has write permissions are
implemented in the same method. Thus an 'unauthorized' error could be
misleading if a user is e.g. space developer but the org is suspended.
With this PR the check for a suspended org is moved into a dedicated
method and all controllers have been adapted to return a different error
in case the user has write permissions but the org is suspended (still
an HTTP 403 error but with a different message).

Co-authored-by: Johannes Haass <johannes.haass@sap.com>
Co-authored-by: Philipp Thun <philipp.thun@sap.com>