Fix flakey test in droplets.spec
The test failed because the generated expected droplet URL is based on the current timestamp. Depending on the test execution speed, it can differ from the timestamp in the URL, which was generated during the request.

Fixed by using Timecop.freeze, which ensures that both URLs will have the same timestamp.