Fix service instance operations duplicates root cause
There was a race condition when two workers were processing two different operations on the same service instance.
The worker first removed the last_operation of that service instance from the service_instance_operations DB table and after that inserted a new operation.
This was done in two seperate transactions, which enabled the second process to insert it's operation after the first process deleted the operation, but before it can insert the new operation. After both processes finished, two operations remained in the table.

This is now prevented by encapsulating the deletion of the operation and the insert of the new operation in a transaction and also locking the service instance entry in the DB until the transaction is finished.