Rename log_rate_limit_per_second to log-rate-limit-per-second
All our archeology seems to indicate that dashes should be preferred in
manifests as they are yaml and yaml prefers kebab-case to snake_case.

We also uncovered some oddities around setting process level properties
on the top level app object. We added tests to document that disk and
memory don't show up in the diff when changed on the app instead of the
process. We also discovered that proprties like health-check that
shouldbe kebab-case were not working at the app level and fixed them to
be consistent.

Signed-off-by: Rebecca Roberts <robertsre@vmware.com>
Co-authored-by: Rebecca Roberts <robertsre@vmware.com>