Remove unnecessary DB checks when creating space roles
Previously we made separate queries to the spaces and organizations tables to check whether the user trying to create a space role was a) able to read the org in question, and b) not able to read the relevant space. In such a case, we sent a 422 Unprocessable Entity response. The effect of this was to serve 422 responses to org auditors and org billing managers. This behaviour is not described in the API docs, and serves no obvious purpose. Org auditorss and org billing managers, like all other roles except admin, org manager and space manager, are unauthorized to create space roles, and should receive a 403 response saying so.