Always allow org managers to assign roles for out-of-org users
This ends an inconsistency whereby org managers received a 201 if they assigned an org role to an unaffiliated user by providing that user's username, but a 422 CF-UnprocessableEntity if they tried to do the same by providing the user's guid. Resolving this inconsistency also allows us to avoid a series of expensive DB queries that were run on all creates where a user guid was provided, which were aimed at determining whether the assignee user was in the same org as the user trying to assign them a role.