Prevent deadlocks when updating managed service instances
Problem
-------
A service instance was in state "update in progress" but the broker did
not receive any request. Retrying the update was not possible
(CF-AsyncServiceInstanceOperationInProgress).

Root cause
----------
Updating the operation (see [1]) and creating the background job ([2])
is done independently. When the latter fails (e.g. database temporarily
not available), the so called UpdaterLock ([3]) is not cleared.

Solution
--------
Introduce similar methods as for ServiceBrokerUpdate ([4]):
- update_broker_needed?
- update_sync
- enqueue_update.
Ensure that operation (lock) and delayed job are handled consistently,
i.e. when the creation of the job fails, the operation will be set to
"failed". This is achieved by moving the creation of the delayed job
into the ServiceInstanceUpdateManaged action (method "enqueue_update"):

  lock = UpdaterLock.new(...)
  lock.lock!

  begin
    update_job = UpdateServiceInstanceJob.new(...)
    pollable_job = Jobs::Enqueuer.new(update_job, ...).enqueue_pollable
    lock.asynchronous_unlock!
  ensure
    lock.unlock_and_fail! if lock.present? && lock.needs_unlock?
  end

[1] https://github.com/cloudfoundry/cloud_controller_ng/blob/05f6cd8f8d3529d5bf8aaccfa2b9c4e75bbfb497/app/controllers/v3/service_instances_controller.rb#L292
[2] https://github.com/cloudfoundry/cloud_controller_ng/blob/05f6cd8f8d3529d5bf8aaccfa2b9c4e75bbfb497/app/controllers/v3/service_instances_controller.rb#L301
[3] https://github.com/cloudfoundry/cloud_controller_ng/blob/main/app/actions/services/locks/updater_lock.rb
[4] https://github.com/cloudfoundry/cloud_controller_ng/blob/main/app/actions/v3/service_broker_update.rb