Return 503 in case of database connection errors
An unhandled Sequel::DatabaseDisconnectError results in a 500 error
response (Internal Server Error). Instead a 503 (Service Unavailable) is
correct as the problem is not caused by internal Cloud Controller
coding.

The DatabaseDisconnectError has been seen on a foundation; the
Sequel::DatabaseConnectionError has been added as it is a similar error
type.