Ignore sequel_pg logs
Before we saw many log entries which had as resource sequel_pg, which is
not very helpful if you want to find the origin of the sql statement.
Ignoring sequel_pg logs in db.rb will skip those entries and make it
easier to see where a sql statement comes from.