Remove unnecessary db query as admin user for organization_quotas
Improve cases for admin user when fetching organization_quotas. Before
this triggered a "SELECT guids FROM organizations" without condition. To
avoid this unnecessary db query for admin users we distinguish now
between admin or not admin and only in the non-admin case the query for
the organization guids is done.