Optimisations for /v3/roles API calls (#2971)
* Remove unnecessary DB checks when creating space roles

Previously we made separate queries to the spaces and organizations tables to check whether the user trying to create a space role was a) able to read the org in question, and b) not able to read the relevant space. In such a case, we sent a 422 Unprocessable Entity response. The effect of this was to serve 422 responses to org auditors and org billing managers. This behaviour is not described in the API docs, and serves no obvious purpose. Org auditorss and org billing managers, like all other roles except admin, org manager and space manager, are unauthorized to create space roles, and should receive a 403 response saying so.

* Always allow org managers to assign roles for out-of-org users

This ends an inconsistency whereby org managers received a 201 if they assigned an org role to an unaffiliated user by providing that user's username, but a 422 CF-UnprocessableEntity if they tried to do the same by providing the user's guid. Resolving this inconsistency also allows us to avoid a series of expensive DB queries that were run on all creates where a user guid was provided, which were aimed at determining whether the assignee user was in the same org as the user trying to assign them a role.

* Improve space_guid and organization_guid methods in role

- Check if a relation is specified (i.e. id != -1) before trying to access them (which results in an unnecessary SELECT).
- Allow getting the organization_guid also for space roles; i.e. the guid of the org which the space belongs to.

* Access organization_guid on space roles when deleting roles

Co-authored-by: Will Gant <william.gant@sap.com>
Co-authored-by: Philipp Thun <philipp.thun@sap.com>