Exclude /healthz from request logging
The logs for requests to /healthz do not contain any useful data. The
user is always empty, the IP address is 127.0.0.1 and for the VCAP
request ID no additinal information can be found as this endpoint does
nothing.

We also have the nginx access logs which contain the exact same
information (including the upstream HTTP status code).