Enable CC to process large manifest
Until now, when a user does a cf push with an app manifest larger than ~130KB, the CC will error with a 500 - unknown error.
This is due to Thin (Puma does the same), which outsources large request bodies/headers into Tempfiles, instead of in-memory StreamIOs.
Currently the CC coding calls a `.string` method, which is not present for Tempfile objects.
This commit changes to the `.read` method, which is available on both classes.
It's also introducing a hard limit of 1MB for the request body size.