Check space/org permissions by id, not guid (#3021)
Prior to this many controllers did a DB query for a space/org to retrieve its guid, which was ultimately passed to VCAP::CloudController::Membership, which would use the guid to do a DB query looking up the exact same thing so its id could be retrieved. In these cases we now retrieve the id in the first query and just pass this through.