Canceling deployment is superseded by new deployment
There are two active states for deployments (DEPLOYING, CANCELING) and
two corresponding actions (DeploymentCreate, DeploymentCancel). This
results in the following possible scenarios:

      |    state    +  action  =  updated state  /  (new deployment)
 ---------------------------------------------------------------------
   1  |  DEPLOYING  |  create  |    SUPERSEDED   |    DEPLOYING
   2  |  DEPLOYING  |  cancel  |    CANCELING    |    -
   3  |  CANCELING  |  cancel  |    CANCELING    |    -
   4  |  CANCELING  |  create  |    CANCELING    |    DEPLOYING

Scenario 4 (CANCELING + create) currently leads to two active
deployments for an app (CANCELING + DEPLOYING). As the Dispatcher and
Updater (module DeploymentUpdater) are not designed to handle two active
deployments concurrently, the result for this scenario has been changed:

      |    state    +  action  =  updated state  /  (new deployment)
 ---------------------------------------------------------------------
   4  |  CANCELING  |  create  |    SUPERSEDED   |    DEPLOYING

This means that the 'create' action now sets a CANCELING deployment to
SUPERSEDED (CANCELED) and during the next run of the DeploymentUpdater,
the new DEPLOYING deployment will be processed.

Furthermore the Updater (module DeploymentUpdater) now checks that the
deployment being processed is still in the expected state.