Finalize only deployments with 'STATUS_VALUE' 'ACTIVE'
In case a deployment of a new or stopped app is triggered the deployments table entry will look like this:
  id                                  | 43
  guid                                | 79165acb-420b-4c46-bc2f-79d5b05b4f0d
  state                               | DEPLOYED
  deploying_web_process_guid          |
  ...
  status_value                        | FINALIZED
  status_reason                       | DEPLOYED
  strategy                            | rolling

As `deploying_web_process_guid` is empty the `finalize_degenerate_deployments` jobs will eventually set the `status_reason` to `DEGENERATE`.
If this happens at a bad moment it can cause a `Cannot cancel a deployment with status: FINALIZED and reason: DEGENERATE` cf push error.
Because the deployment is already `FINALIZED` it does not make sense to change its status.
Therefore this change makes sure that `finalize_degenerate_deployments` only considers active deployments.

Co-authored-by: Philipp Thun <philipp.thun@sap.com>