Skip interim web processes from SUPERSEDED (CANCELED) deployments
When a deployment train is canceled, the latest interim web process is
scaled up. The introduction of a SUPERSEDED (CANCELED) state requires a
change to how this latest interim web process is determined.

Let's assume the following sequence of actions:
  create
  create
  cancel
  create
  cancel

This would result in the following states (after each action):
  DEPLOYING
  SUPERSEDED (DEPLOYED) + DEPLOYING
  SUPERSEDED (DEPLOYED) + CANCELING
  SUPERSEDED (DEPLOYED) + SUPERSEDED (CANCELED) + DEPLOYING
  SUPERSEDED (DEPLOYED) + SUPERSEDED (CANCELED) + CANCELING

And this should result in the web process from the SUPERSEDED (CANCELED)
deployment to be skipped and the one from the SUPERSEDED (DEPLOYED)
deployment to be scaled up.