Scale canceled web processes to zero
The introduction of a SUPERSEDED (CANCELED) state requires the Updater
(module DeploymentUpdater) to handle the interim web process belonging
to such a SUPERSEDED (CANCELED) deployment.

The cleanup of unused web processes happens at the end of a (successful
or canceled) deployment. But as the canceled interim web process should
not be used anymore, it is now immediately scaled to zero instances.