Truncate to 2**13 for service broker errors
- so the previous commit got us close, but we were still not writing the
error at all.  the user would only see null if the error was too big
- I still don't understand all the flows of who tries to do what with
cf_api_error, but changing the truncation from 2**14 to 2**13 makes the
broker error small enough to correctly fit in all the tables
- This seems fine since it fits in the smaller cf_api_error and no other
issues going on, but may continue not to be the best place to put
truncation.  But is probably fine for our service instance errors
problem
- reverted other changes to make the smallest changeset possible that
seems to behave

Authored-by: Michael Oleske <moleske@pivotal.io>