Canceling deployment is superseded by new deployment (#3072)
* Canceling deployment is superseded by new deployment

There are two active states for deployments (DEPLOYING, CANCELING) and
two corresponding actions (DeploymentCreate, DeploymentCancel). This
results in the following possible scenarios:

      |    state    +  action  =  updated state  /  (new deployment)
 ---------------------------------------------------------------------
   1  |  DEPLOYING  |  create  |    SUPERSEDED   |    DEPLOYING
   2  |  DEPLOYING  |  cancel  |    CANCELING    |    -
   3  |  CANCELING  |  cancel  |    CANCELING    |    -
   4  |  CANCELING  |  create  |    CANCELING    |    DEPLOYING

Scenario 4 (CANCELING + create) currently leads to two active
deployments for an app (CANCELING + DEPLOYING). As the Dispatcher and
Updater (module DeploymentUpdater) are not designed to handle two active
deployments concurrently, the result for this scenario has been changed:

      |    state    +  action  =  updated state  /  (new deployment)
 ---------------------------------------------------------------------
   4  |  CANCELING  |  create  |    SUPERSEDED   |    DEPLOYING

This means that the 'create' action now sets a CANCELING deployment to
SUPERSEDED (CANCELED) and during the next run of the DeploymentUpdater,
the new DEPLOYING deployment will be processed.

Furthermore the Updater (module DeploymentUpdater) now checks that the
deployment being processed is still in the expected state.

* Skip interim web processes from SUPERSEDED (CANCELED) deployments

When a deployment train is canceled, the latest interim web process is
scaled up. The introduction of a SUPERSEDED (CANCELED) state requires a
change to how this latest interim web process is determined.

Let's assume the following sequence of actions:
  create
  create
  cancel
  create
  cancel

This would result in the following states (after each action):
  DEPLOYING
  SUPERSEDED (DEPLOYED) + DEPLOYING
  SUPERSEDED (DEPLOYED) + CANCELING
  SUPERSEDED (DEPLOYED) + SUPERSEDED (CANCELED) + DEPLOYING
  SUPERSEDED (DEPLOYED) + SUPERSEDED (CANCELED) + CANCELING

And this should result in the web process from the SUPERSEDED (CANCELED)
deployment to be skipped and the one from the SUPERSEDED (DEPLOYED)
deployment to be scaled up.

* Scale canceled web processes to zero

The introduction of a SUPERSEDED (CANCELED) state requires the Updater
(module DeploymentUpdater) to handle the interim web process belonging
to such a SUPERSEDED (CANCELED) deployment.

The cleanup of unused web processes happens at the end of a (successful
or canceled) deployment. But as the canceled interim web process should
not be used anymore, it is now immediately scaled to zero instances.