Patch HTTPClient to accept socket_connect_timeout
Monkey patch HTTPClient to accept socket_connect_timeout as optional
configuration. If set, it is used as connect_timeout parameter when
creating a TCPSocket.

Set the socket_connect_timeout in Diego::Client to (overall)
connect_timeout / 2.

Add test to ensure that the httpclient version is not updated without
revisiting the monkey patch.

Co-authored-by: Johannes Haass <johannes.haass@sap.com>