Ensure that delayed_job workers handle `SIGQUIT`
Adds a monkey patches to delayed_job workers to gracefully shutdown in case of a `QUIT` signal.
So far delayed_job only considered `INT` & `TERM` signals for graceful shutdown.
This might become obsolete once the following change is merged into delayed_job.
See https://github.com/collectiveidea/delayed_job/pull/900

To ensure that a version bump of delayed_job does not break the monkey patch
a test will fail in case the version number changes.
In case of a version bump it needs to be manually verified if the monkey
patch still works.

Co-authored-by: Philipp Thun <philipp.thun@sap.com>
Co-authored-by: Sven Krieger <37476281+svkrieger@users.noreply.github.com>
Co-authored-by: Johannes Haass <johannes.haass@sap.com>