Keep original VCAP-Request-ID for re-enqueued ReoccurringJobs
All background jobs are wrapped in a LoggingContextJob that stores the
original VCAP-Request-ID and ensures that the same ID is used (e.g. for
logs) while performing the job.

For ReoccurringJobs, this capability got lost when the job was
re-enqueued. Adding the next job to the queue did not happen as part of
the 'perform' method, but in the 'success' hook. So by using the
'with_request_id_set' method also for the 'success' hock, the
re-enqueued job keeps the original VCAP-Request-ID.

The effect can be seen when looking at logs with source
cc.service_broker.v2.http_client. When polling for the last operation
state of a provisioned service instance, the VCAP-Request-ID was only
set for the initial 'fetch' request (which was done in the same job as
the 'provision' request). With this change, all subsequent state polls
are also logged with "data":{"request_guid":"original-guid",...}.