Compare only spec.destinations for route_mappings [unit test]
When indexes are created for route_mappings, this unit test
fails for postgres 10 and 11 version, because it compare the whole hash
and the order of the spec.destinations array was swapped for these two
versions. Nevertheless, we are intreseted in the spec.destinations array
and this changes match that exactly.