Add validation for resource match mode
If provided, the 'mode' field (resource match object) must be a string.
Without this fix a request to /v3/resource_matches with

  { .. "mode": 644 .. }

resulted in an UnknownError. Now the error message

  "array contains at least one resource with a non-string mode"

is returned instead.

Furthermore the format is checked and if the provided value is not a
POSIX mode string (*), the following error message is returned:

  "array contains at least one resource with an incorrect mode"

(*) A simple regex only allowing digits from 0 to 7 is used. At least 1
digit must be provided.

Co-authored-by: Florian Braun <5863788+FloThinksPi@users.noreply.github.com>