Add delayed_jobs_reserve_where index
With PR #3324 [1] the 'delayed_jobs_reserve' index was changed by adding the field 'priority' which is used in the ORDER BY clause. This change re-adds the previous index, that only contains fields used in the WHERE clause of the query. Although PostgreSQL could always use the new index, there seem to be situations where the query planner decides for a sequential table scan (theory: if the number of entries is rather low).

[1] https://github.com/cloudfoundry/cloud_controller_ng/pull/3324

Co-authored-by: Dimitar Velinov <dimitar.velinov01@sap.com>