Improve SQL Performance by Eliminating Unnecessary Order
This commit corrects and improves SQL queries by avoiding redundant orders. It was previously observed that SQL queries were being ordered both by ID and GUID when the default order was applied. The paginator was also adding a secondary sort, introduced to offset issues encountered when using 'created_at' or 'updated_at' as sorting parameters, as the order was not deterministic. However, for default sorting by ID, this was superfluous.

With this change, the paginator now adds the secondary sort to the SQL query only if the order by parameter is not the ID column, in order to ensure fixed order for all column values. This alteration aims to enhance query performance across all listing endpoints.

The change involves a schema update that aligns all related indices in the database. A combined index with the GUID column is created with the understanding that databases can use such combined indices as a replacement for the most left column in the index. For instance, an index over (updated_at, GUID) can be utilized by the database, providing enhanced performance either when querying this table with both fields in the where clause, or when ordering by both columns.

The primary and secondary ordering now use the same direction, an improvement from the previous method where order was only applied to primary sorting. This prevents the need for separate indices for ascending and descending orders, resulting in a system where descending order is the exact opposite of ascending order.

Co-authored-by: Johannes Haass <johannes.haass@sap.com>